+++ src/operating_systems/netbsd/qemu_vm.ts
@@ -8,4 +8,8 @@ export class Vm extends QemuVm {
   protected override async shutdown(): Promise<void> {
     await this.execute('sudo shutdown -h -p now')
   }
+
+  protected override get ipv6(): string {
+    return 'ipv6=off'
+  }
 }