+++ changelog.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for configuring CPU core count ([#16](https://github.com/cross-platform-actions/action/issues/17))
 
 ### Changed
+- Remove support for IPv6 for NetBSD ([#46](https://github.com/cross-platform-actions/action/issues/46))
 - Increased default memory to 13GB on macOS runner and to 6GB on Linux runners ([#16](https://github.com/cross-platform-actions/action/issues/16))
 - Increased default CPU core count to 3 on macOS runner and to 2 on Linux runners ([#16](https://github.com/cross-platform-actions/action/issues/17))
 - Changed from two CPU sockets to one CPU socket ([#16](https://github.com/cross-platform-actions/action/issues/17))
 
 ### Fixed
+- NetBSD - very slow network ([#46](https://github.com/cross-platform-actions/action/issues/46))
 - Action doesn't terminate when command fails ([#21](https://github.com/cross-platform-actions/action/issues/21))
 
 ## [0.10.0] - 2023-01-24